requirements.txt
openai
PyYAML
backoff
sentence-transformers
bs4
